Journal: add mustache template, new journal layout
Closes #2467
