+++ pkg/shell/shell.html
@@ -6,6 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="../base1/cockpit.css" rel="stylesheet">
     <link href="shell.css" rel="stylesheet">
+    <link href="../system/journal.css" rel="stylesheet">
     <script src="../base1/bundle.js"></script>
     <script src="../shell/bundle.js"></script>
     <script src="../system/bundle.js"></script>
