+++ pkg/systemd/Makefile.am
@@ -2,9 +2,9 @@ systemddir = $(pkgdatadir)/system
 systemd_DATA = \
 	pkg/systemd/bundle.min.js.gz \
 	pkg/systemd/init.min.html.gz \
-	pkg/systemd/log.min.html.gz \
 	pkg/systemd/manifest.json \
 	pkg/systemd/terminal.min.html.gz \
+	pkg/systemd/journal.min.html.gz \
 	$(NULL)
 
 systemddebugdir = $(DBGDIR)$(systemddir)
@@ -12,18 +12,29 @@ systemddebug_DATA = \
 	pkg/systemd/bundle.js \
 	pkg/systemd/init.js \
 	pkg/systemd/init.html \
-	pkg/systemd/log.html \
-	pkg/systemd/log.js \
 	pkg/systemd/server.js \
 	pkg/systemd/terminal.js \
 	pkg/systemd/terminal.html \
+	pkg/systemd/journal.html \
+	pkg/systemd/journal.js \
+	pkg/systemd/journal.css \
+	pkg/systemd/journalctl.js \
+	pkg/systemd/renderer.js \
+	pkg/systemd/journal_day_header.mustache \
+	pkg/systemd/journal_line.mustache \
+	pkg/systemd/journal_reboot.mustache \
 	$(NULL)
 
 systemd_BUNDLE = \
 	pkg/systemd/init.min.js \
-	pkg/systemd/log.min.js \
 	pkg/systemd/server.min.js \
 	pkg/systemd/terminal.min.js \
+	pkg/systemd/journal.min.js \
+	pkg/systemd/journalctl.min.js \
+	pkg/systemd/renderer.min.js \
+	pkg/systemd/journal_day_header.mustache \
+	pkg/systemd/journal_line.mustache \
+	pkg/systemd/journal_reboot.mustache \
 	$(NULL)
 
 pkg/systemd/bundle.min.js: $(systemd_BUNDLE)
@@ -38,16 +49,16 @@ TESTS += $(systemd_TESTS)
 CLEANFILES += \
 	pkg/systemd/bundle.min.js \
 	pkg/systemd/init.min.html \
-	pkg/systemd/log.min.html \
 	pkg/systemd/terminal.min.html \
+	pkg/systemd/journal.min.html \
 	$(systemd_BUNDLE) \
 	$(NULL)
 
 EXTRA_DIST += \
 	pkg/systemd/bundle.min.js \
 	pkg/systemd/init.min.html \
-	pkg/systemd/log.min.html \
 	pkg/systemd/terminal.min.html \
+	pkg/systemd/journal.min.html \
 	$(systemd_DATA) \
 	$(systemddebug_DATA) \
 	$(systemd_TESTS) \
