+++ pkg/systemd/init.html
@@ -5,6 +5,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="../base1/cockpit.css" type="text/css" rel="stylesheet">
+    <link href="journal.css" type="text/css" rel="stylesheet">
     <script src="../base1/bundle.js"></script>
     <script src="../system/bundle.js"></script>
 <script>
