+++ pkg/systemd/journal_day_header.mustache
@@ -1,0 +1,1 @@
+<div class="panel-heading">{{day}}</div>
