+++ pkg/systemd/journal_reboot.mustache
@@ -1,0 +1,5 @@
+<div class="cockpit-logline">
+  {{! placeholders for correct message alignment }}
+  <div class="cockpit-log-warning"></div>
+  <span class="cockpit-log-message cockpit-logmsg-reboot">{{message}}</span>
+</div>
