+++ pkg/systemd/manifest.json
@@ -7,7 +7,7 @@
             "label": "Services",
             "order": 5
         },
-        "log": {
+        "journal": {
             "label": "Journal",
             "order": 8
         }
