+++ po/POTFILES.in
@@ -1,7 +1,7 @@
 [encoding: UTF-8]
 # List of source files containing translatable strings.
 # Please keep this file sorted alphabetically.
-pkg/systemd/log.html
+pkg/systemd/journal.html
 pkg/systemd/server.js
 [type: gettext/glade] pkg/shell/shell.html
 pkg/shell/cockpit-cpu-status.js
